Further revision of the JsonConnection
Includes new JsonClientHandler that uses the new JsonConnection’s
ObjectMapper.
